+++ snapatac2-core/src/preprocessing/count_data.rs
@@ -5,7 +5,7 @@ mod matrix;
 
 pub use crate::preprocessing::qc;
 pub use import::{import_fragments, import_contacts};
-pub use coverage::{GenomeCoverage, ContactMap, CoverageType, fragments_to_insertions};
+pub use coverage::{GenomeCount, ContactMap, FragmentType, fragments_to_insertions};
 pub use genome::{
     Transcript, Promoters, FeatureCounter, TranscriptCount, GeneCount,
     read_transcripts_from_gff, read_transcripts_from_gtf,
@@ -45,7 +45,7 @@ pub trait SnapData: AnnDataOp {
 
     /// Read insertion counts stored in the `.obsm` matrix.
     fn get_count_iter(&self, chunk_size: usize) ->
-        Result<GenomeCoverage<Box<dyn ExactSizeIterator<Item = (CoverageType, usize, usize)>>>>;
+        Result<GenomeCount<Box<dyn ExactSizeIterator<Item = (FragmentType, usize, usize)>>>>;
 
     fn contact_count_iter(&self, chunk_size: usize) -> Result<ContactMap<Self::CountIter>>;
 
@@ -72,7 +72,7 @@ pub trait SnapData: AnnDataOp {
 
     /// Compute TSS enrichment.
     fn tss_enrichment(&self, promoter: &BedTree<bool>) -> Result<Vec<f64>> {
-        Ok(self.get_count_iter(2000)?.into_raw().flat_map(|(fragments, _, _)| {
+        Ok(self.get_count_iter(2000)?.into_fragments().flat_map(|(fragments, _, _)| {
             fragments.into_par_iter()
                 .map(|x| qc::tss_enrichment(x.into_iter(), promoter))
                 .collect::<Vec<_>>()
@@ -84,7 +84,7 @@ pub trait SnapData: AnnDataOp {
 
     /// Compute the fraction of reads in each region.
     fn frip<D>(&self, regions: &Vec<BedTree<D>>) -> Result<Array2<f64>> {
-        let vec = qc::fraction_of_reads_in_region(self.get_count_iter(2000)?.into_raw(), regions)
+        let vec = qc::fraction_of_reads_in_region(self.get_count_iter(2000)?.into_fragments(), regions)
             .map(|x| x.0).flatten().flatten().collect::<Vec<_>>();
         Array2::from_shape_vec((self.n_obs(), regions.len()), vec).map_err(Into::into)
     }
@@ -94,18 +94,18 @@ impl<B: Backend> SnapData for AnnData<B> {
     type CountIter = ChunkedArrayElem<B, CsrMatrix<u8>>;
 
     fn get_count_iter(&self, chunk_size: usize) ->
-        Result<GenomeCoverage<Box<dyn ExactSizeIterator<Item = (CoverageType, usize, usize)>>>>
+        Result<GenomeCount<Box<dyn ExactSizeIterator<Item = (FragmentType, usize, usize)>>>>
     {
         let obsm = self.obsm();
-        let matrices: Box<dyn ExactSizeIterator<Item = (CoverageType, usize, usize)>> =
+        let matrices: Box<dyn ExactSizeIterator<Item = (FragmentType, usize, usize)>> =
             if let Some(insertion) = obsm.get_item_iter("fragment_single", chunk_size) {
-                Box::new(insertion.map(|(x, a, b)| (CoverageType::FragmentSingle(x), a, b)))
+                Box::new(insertion.map(|(x, a, b)| (FragmentType::FragmentSingle(x), a, b)))
             } else if let Some(fragment) = obsm.get_item_iter("fragment_paired", chunk_size) {
-                Box::new(fragment.map(|(x, a, b)| (CoverageType::FragmentPaired(x), a, b)))
+                Box::new(fragment.map(|(x, a, b)| (FragmentType::FragmentPaired(x), a, b)))
             } else {
                 anyhow::bail!("neither 'fragment_single' nor 'fragment_paired' is present in the '.obsm'")
             };
-        Ok(GenomeCoverage::new(self.read_chrom_sizes()?, matrices))
+        Ok(GenomeCount::new(self.read_chrom_sizes()?, matrices))
     }
 
     fn contact_count_iter(&self, chunk_size: usize) -> Result<ContactMap<Self::CountIter>> {
@@ -128,19 +128,19 @@ impl<B: Backend> SnapData for AnnDataSet<B> {
     type CountIter = StackedChunkedArrayElem<B, CsrMatrix<u8>>;
 
     fn get_count_iter(&self, chunk_size: usize) ->
-        Result<GenomeCoverage<Box<dyn ExactSizeIterator<Item = (CoverageType, usize, usize)>>>>
+        Result<GenomeCount<Box<dyn ExactSizeIterator<Item = (FragmentType, usize, usize)>>>>
     {
         let adatas = self.adatas().inner();
         let obsm = adatas.get_obsm();
-        let matrices: Box<dyn ExactSizeIterator<Item = (CoverageType, usize, usize)>> =
+        let matrices: Box<dyn ExactSizeIterator<Item = (FragmentType, usize, usize)>> =
             if let Some(insertion) = obsm.get_item_iter("fragment_single", chunk_size) {
-                Box::new(insertion.map(|(x, a, b)| (CoverageType::FragmentSingle(x), a, b)))
+                Box::new(insertion.map(|(x, a, b)| (FragmentType::FragmentSingle(x), a, b)))
             } else if let Some(fragment) = obsm.get_item_iter("fragment_paired", chunk_size) {
-                Box::new(fragment.map(|(x, a, b)| (CoverageType::FragmentPaired(x), a, b)))
+                Box::new(fragment.map(|(x, a, b)| (FragmentType::FragmentPaired(x), a, b)))
             } else {
                 anyhow::bail!("neither 'fragment_single' nor 'fragment_paired' is present in the '.obsm'")
             };
-        Ok(GenomeCoverage::new(self.read_chrom_sizes()?, matrices))
+        Ok(GenomeCount::new(self.read_chrom_sizes()?, matrices))
     }
 
     fn contact_count_iter(&self, chunk_size: usize) -> Result<ContactMap<Self::CountIter>> {