+++ snapatac2-core/src/preprocessing/count_data/matrix.rs
@@ -15,12 +15,20 @@ use bed_utils::bed::{BEDLike, tree::{GenomeRegions, SparseCoverage}};
 /// 
 /// # Arguments
 /// 
-/// * `anndata` - 
+/// * `adata` - The input anndata object.
+/// * `bin_size` - The bin size.
+/// * `chunk_size` - The chunk size.
+/// * `exclude_chroms` - The chromosomes to exclude.
+/// * `min_fragment_size` - The minimum fragment size.
+/// * `max_fragment_size` - The maximum fragment size.
+/// * `out` - The output anndata object.
 pub fn create_tile_matrix<A, B>(
     adata: &A,
     bin_size: usize,
     chunk_size: usize,
     exclude_chroms: Option<&[&str]>,
+    min_fragment_size: Option<u64>,
+    max_fragment_size: Option<u64>,
     out: Option<&B>,
     ) -> Result<()>
 where
@@ -46,11 +54,19 @@ where
         }
     } else {
         let mut counts = adata.get_count_iter(chunk_size)?.with_resolution(bin_size);
+
         if let Some(exclude_chroms) = exclude_chroms {
             counts = counts.exclude(exclude_chroms);
         }
+        if let Some(min_fragment_size) = min_fragment_size {
+            counts = counts.min_fragment_size(min_fragment_size);
+        }
+        if let Some(max_fragment_size) = max_fragment_size {
+            counts = counts.max_fragment_size(max_fragment_size);
+        }
+
         let feature_names = counts.get_gindex().to_index().into();
-        let data_iter = counts.into_values::<u32>().map(|x| x.0).progress_with_style(style);
+        let data_iter = counts.into_counts::<u32>().map(|x| x.0).progress_with_style(style);
         if let Some(adata_out) =  out {
             adata_out.set_x_from_iter(data_iter)?;
             adata_out.set_obs_names(adata.obs_names())?;
@@ -67,6 +83,8 @@ pub fn create_peak_matrix<A, I, D, B>(
     adata: &A,
     peaks: I,
     chunk_size: usize,
+    min_fragment_size: Option<u64>,
+    max_fragment_size: Option<u64>,
     out: Option<&B>,
     use_x: bool,
     ) -> Result<()>
@@ -86,8 +104,14 @@ where
         Box::new(adata.read_chrom_values(chunk_size)?
             .aggregate_by(counter).map(|x| x.0))
     } else {
-        Box::new(adata.get_count_iter(chunk_size)?
-            .aggregate_by(counter).map(|x| x.0))
+        let mut counts = adata.get_count_iter(chunk_size)?;
+        if let Some(min_fragment_size) = min_fragment_size {
+            counts = counts.min_fragment_size(min_fragment_size);
+        }
+        if let Some(max_fragment_size) = max_fragment_size {
+            counts = counts.max_fragment_size(max_fragment_size);
+        }
+        Box::new(counts.aggregate_counts_by(counter).map(|x| x.0))
     };
     if let Some(adata_out) =  out {
         adata_out.set_x_from_iter(data.progress_with_style(style))?;
@@ -105,6 +129,8 @@ pub fn create_gene_matrix<A, B>(
     transcripts: Vec<Transcript>,
     id_type: &str, 
     chunk_size: usize,
+    min_fragment_size: Option<u64>,
+    max_fragment_size: Option<u64>,
     out: Option<&B>,
     use_x: bool,
     ) -> Result<()>
@@ -122,8 +148,14 @@ where
                 Box::new(adata.read_chrom_values(chunk_size)?
                     .aggregate_by(transcript_counter).map(|x| x.0))
             } else {
-                Box::new(adata.get_count_iter(chunk_size)?
-                    .aggregate_by(transcript_counter).map(|x| x.0))
+                let mut counts = adata.get_count_iter(chunk_size)?;
+                if let Some(min_fragment_size) = min_fragment_size {
+                    counts = counts.min_fragment_size(min_fragment_size);
+                }
+                if let Some(max_fragment_size) = max_fragment_size {
+                    counts = counts.max_fragment_size(max_fragment_size);
+                }
+                Box::new(counts.aggregate_counts_by(transcript_counter).map(|x| x.0))
             };
             if let Some(adata_out) = out {
                 adata_out.set_x_from_iter(data)?;
@@ -143,8 +175,14 @@ where
                 Box::new(adata.read_chrom_values(chunk_size)?
                     .aggregate_by(gene_counter).map(|x| x.0))
             } else {
-                Box::new(adata.get_count_iter(chunk_size)?
-                    .aggregate_by(gene_counter).map(|x| x.0))
+                let mut counts = adata.get_count_iter(chunk_size)?;
+                if let Some(min_fragment_size) = min_fragment_size {
+                    counts = counts.min_fragment_size(min_fragment_size);
+                }
+                if let Some(max_fragment_size) = max_fragment_size {
+                    counts = counts.max_fragment_size(max_fragment_size);
+                }
+                Box::new(counts.aggregate_counts_by(gene_counter).map(|x| x.0))
             };
             if let Some(adata_out) = out {
                 adata_out.set_x_from_iter(data)?;